move pending map to a new file
